+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-Cambio V.1.1.06 - ✨ Últimas mejoras implementadas:
+Cambio V.1.1.08 - ✨ Últimas mejoras implementadas:
+
+- Añadido editor de imagenes en el editor de texto. Ahora cuando se va a adjuntar una imagen, te permitira editarla antes de embedirla. Las acciones que perimte son: Recortar, marcar, anotar y flecha.
+- Añadido corrector ortografico en el editor de texto. Se debe de instalar la libreria pyspellchecker. Leer el archivo README.md para poder instalarlo. En caso de no instalarlo, el editor funciona pero sin esa funcion.
+En ajustes de usuario, en la pestaña General, se puede cambiar el idioma.
+- En el historial de los expedientes, ya se permite seleccionar texto para las funciones de copiar, cortar y pegar.
+- Todo documentado en ayuda de la app.
+
+
+Cambio V.1.1.06
 
 - Añadido a la ventana de asociaciones de la ficha de los expedientes, el poder importar emails y asi asociarlos al expediente.
 El email puede estar en formato .eml y .msg